requirements.txt
fastapi
uvicorn
streamlit
langgraph
langchain-core
langchain-community
chromadb
pandas
python-multipart
google-generativeai
python-dotenv
sqlalchemy
pydantic
requests
pypdf
cryptography


